feat(main): just enough to bakelite itself
This is just the beginning of the amazing adventures of Terin and
Bakelite. Their journey is destined to be packed with non-stop action,
millions of laughs, heart pounding perils, and endless excitement.
Together, they'll encounter fantastic friends, evil enemies, and meet
creatures beyond your wildest imagination. And as their story unfolds,
we'll unlock the magic and mystery of a most wondrous place, the
incredible world of [fixme].