feat(main): support modifying built platforms
Bakelite now accepts a platform modification string vita the -platform
flag. This string format for adding and removing the built platforms,
including removing the default platforms entirely. It is processed left
to right.

  bakelite -platform '-windows +linux/s390x' [packages]

Would now build [packages] for all of the default platforms, except for
Windows, and also build for the linux/s390x platform.

  bakelite -platform '- +linux +darwin' [packages]

Would now build [packages] for the default platforms of only the Linux
and Darwin operating systems.

The platform modification string is parsed and turned into calls to
a structure following the builder pattern.

As the user can now modify the built platforms, Bakelite now exits with
a non-zero exit code if any of the platforms fail to build.

Change-Id: Iade51e17bbfda4e916394343a5f8cb3208f2b160